Added TC for MTA-512 Verify plugin details links
